requirements.txt
click==7.1.2
durable-rules==2.0.28
fastapi==0.58.0
h11==0.9.0
httptools==0.1.1
pkg-resources==0.0.0
pydantic==1.5.1
starlette==0.13.4
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
